requirements.txt
google-generativeai>=0.8.0
openai>=1.40.0
pydantic>=2.0
gradio>=4.0
matplotlib>=3.8
python-dotenv>=1.0
